refactor(main): split the IPC registrar by domain, and drop the roles nothing sends
`ipcHandlers.ts` was 445 lines: 34 handlers and the helpers they lean on,
so a change to the file-open flow and a change to the recent list touched
the same file. Each domain now owns a module under `handlers/` -- app
paths, file dialogs, file system, app state, menu state, recent files,
context menus, window actions -- and the registrar is the list of them,
which puts the set of channels the main process answers on one screen.

`handleInvoke` moves out first, to `ipc/handleInvoke.ts`. Everything that
registers a channel needs those nine lines of type plumbing, so reaching
them through the registrar had `renderWindowIpc` and `cuemolClipboard`
importing the thing that registers everything else.

Two bugs fell out of reading `MENU_INVOKE_ROLE` closely enough to move it:

  - It answered ten roles. Two are declared as `role:` anywhere in the
    menu template; the edit roles resolve by focus and the macOS app-menu
    roles are native. Reload, force-reload, the three zoom roles,
    fullscreen, about and close were never attached to a menu item. The
    zoom ones were the convincing part: they drove an app-wide
    `setZoomLevel` that the Rendering window still adopts when it opens,
    so the feature looked finished from both ends while nothing could
    trigger it.
  - It acted on the main window whatever asked, so DevTools requested
    from the Rendering window opened the main window's inspector. It now
    resolves the sender's window, as the neighbouring handlers already do.
